requirements.txt
Flask==1.1.1
python-aiml==0.9.3